Use larger data type to avoid overflow
uint16_t was used to get scratch space offset. Largest value that can be
encoded for this type is 65536. Each scratch slot using HWord message is
32 bytes. So uint16_t usage is correct only up to 2048 HWord slots. When later sHWord slots are accessed, uint16_t goes OOB. This causes dependency check to become incorrect.

Fix is to use larger size (uint32_t) that can never go OOB.

(cherry picked from commit 10f146fe24343b0ca2a2ad23ea512c393a00bfda)

+++ visa/LocalScheduler/Dependencies_G4IR.cpp
@@ -140,15 +140,15 @@ static DepType DoMemoryInterfereScratchSend(G4_INST *send1, G4_INST *send2,
         (depT == RET_WAW && !send1IsRead && !send2IsRead) ||
         (depT == RET_RAW && !send1IsRead && send2IsRead)) {
       // scratch guaranteed to return valid linear ImmOff
-      uint16_t leftOff1 = (uint16_t)send1->getMsgDesc()->getOffset()->immOff;
-      uint16_t leftOff2 = (uint16_t)send2->getMsgDesc()->getOffset()->immOff;
+      uint32_t leftOff1 = (uint32_t)send1->getMsgDesc()->getOffset()->immOff;
+      uint32_t leftOff2 = (uint32_t)send2->getMsgDesc()->getOffset()->immOff;
       auto bytesAccessed = [](const G4_INST *send) {
         return send->getMsgDesc()->isRead()
                    ? (uint16_t)send->getMsgDesc()->getDstLenBytes()
                    : (uint16_t)send->getMsgDesc()->getSrc1LenBytes();
       };
-      uint16_t rightOff1 = leftOff1 + bytesAccessed(send1) - 1;
-      uint16_t rightOff2 = leftOff2 + bytesAccessed(send2) - 1;
+      uint32_t rightOff1 = leftOff1 + bytesAccessed(send1) - 1;
+      uint32_t rightOff2 = leftOff2 + bytesAccessed(send2) - 1;
       if (leftOff1 > rightOff2 || leftOff2 > rightOff1) {
         return NODEP;
       }
